server/requirements.txt
Flask==2.2.2
pandas==1.5.3
python_dateutil==2.8.2
requests==2.28.1
scikit_learn==1.2.2
